requirements.txt
setuptools
requests~=2.31.0
psycopg2~=2.9.9
sqlalchemy~=2.0.28
progressbar2
numpy
pandas~=2.2.1
geopandas~=0.14.3
